siptrack_migration/requirement.txt
Django==1.11.29
MySQL-python==1.2.5
argparse==1.2.1
certifi==2020.11.8
djangorestframework==3.9.4
pytz==2020.4
urllib3==1.26.2
wsgiref==0.1.2